requirements.txt
pymupdf
python-docx
endee
fastembed
langchain-google-genai
langchain-core
langchain
python-dotenv
fastapi
uvicorn[standard]
python-multipart
